Fixed bug

+++ fk-compute/cpp/src/linalg.cpp
@@ -5,7 +5,7 @@
 
 int computeDotProduct(const std::vector<int> &a, const std::vector<int> &b) {
   int accumulator = a[0];
-  for (int z = 0; z < b.size(); z++) {
+  for (size_t z = 0; z < b.size(); z++) {
     accumulator += a[z + 1] * b[z];
   }
   return accumulator;
